Lock twiddling
Trying to avoid locking in the IO loop. Since changes to the list of sockets is something that happens rarely we can make a copy of the list when we add something, which means that we don't need to guard the access of the list in the loop. We only need to lock over the copy and reassignment to the local variable, so that two threads can add sockets at the same time (and if the loop misses one of the additions it doesn't matter, it will pick it up on the next tick).