Work around a test failure in RBX
This must be an RBX bug. It complained about yielding an arg to a block with arity zero. What actually happened was that a block was captured (query_result.rb#60) and forwarded to another call. Changing that to an explicit #call, or doing what I did in this patch made the error go away.