Rake test phase now starts/stops a 1-node cluster for RSpec tests
This is a quick-and-dirty version.  A more robust one would require to
extract CCM Module from features/env.rb in a lib/ccm.rb file and then
include it in Rakefile.  That way, Cassandra version would not be
duplicated, and all interactions with CCM would be kept in the same file.