Improved execution time of Cucumber tests
Cluster configuration is retrieved from nodetool now instead of being
stored as instance variables.  That way, the first Cucumber test can
leverage an existing cluster if it exists and has the right
configuration.

At the end of authentication tests, the cluster is configured back to an
AllowAllAuthenticator.  This is to make sure that future tests can reuse
the existing cluster even if they assume that authentication is
disabled.

Finally, we make sure that *all* nodes of a cluster are up before
running the test.  Stories can include some "And node 3 stops", it does
not prevent the next stories to run on the same cluster without having
to restart the whole cluster.

On my machine, there improvements reduced the run time from ~11 minutes
to ~7 minutes.